bugfix: (1) separate protein prep from sams sampling; (2) circular mean and circular std.
